Stage 5 Alpha 4

+++ freports/views/courts.py
@@ -8,13 +8,24 @@ from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import ObjectDoesNotExist
 
-from ..models import Court, Judge
+from ..models import Court, Judge, Report
 
 @login_required(login_url='/login/')
 def courts_list(request):
     courts = Court.objects.all().order_by('name')
     header = 'Список судів'
     return render(request, 'freports/courts_list.html', {'courts': courts, 'header': header})
+
+@login_required(login_url='/login/')
+def court_detail(request, cid):
+    court = Court.objects.get(pk=cid)
+    judges = Judge.objects.filter(court_name=court)
+    cases_amount = 0
+    for judge in judges:
+        cases_amount += Report.objects.filter(judge_name=judge).count()
+    header = u'Детальна інформація про {}'.format(court.name)
+    return render(request, 'freports/court_detail.html', {'content': court, 'header': header, 'judges': judges,
+        'cases_amount': cases_amount})
 
 @login_required(login_url='/login/')
 def add_court(request):
@@ -39,20 +50,68 @@ def add_court(request):
     else:
         return render(request, 'freports/court_form.html', {'header': header})
 
+@login_required(login_url='/login/')
+def edit_court(request, cid):
+    court = Court.objects.get(pk=cid)
+    header = u'Редагування інформації про {}'.format(court.name)
+    judges = Judge.objects.filter(court_name=court)
+    if request.method == 'POST':
+        next_url = request.POST.get('next_url', '')
+        if request.POST.get('save_button'):
+            valid_data = valid_court(request.POST)
+            errors = valid_data['errors']
+            new_court = valid_data['new_court']
+            if errors:
+                messages.error(request, u"Виправте наступні помилки")
+                return render(request, 'freports/court_form.html', {'content': new_court, 'errors': errors, 'header': header,
+                    'judges': judges, 'next_url': next_url})
+            else:
+                new_item = Court(**new_court)
+                new_item.id=cid
+                new_item.save()
+                messages.success(request, u"Інформація про %s успішно змінена" % new_item)
+        elif request.POST.get('cancel_button'):
+            messages.warning(request, u"Редагування інформації про {} скасовано".format(court.name))
+        if next_url == '':
+            next_url = reverse('courts_list')
+
+        return HttpResponseRedirect(next_url)
+
+    else:
+        next_url_name, next_url = request.GET.get('next', ''), ''
+        if next_url_name:
+            next_url = reverse(next_url_name, args=[cid])
+        return render(request, 'freports/court_form.html', {'header': header, 'content': court, 'judges': judges,
+            'next_url':next_url})
+
 
 def delete_court(request, cid):
     court = Court.objects.get(pk=cid)
     header = u"Видалення інформації про %s" % court
+    content = u"Ви дійсно бажаєте видалити інформацію про {}?".format(court.name)
     if request.method == "GET":
-        return render(request, 'freports/delete_form.html', {'content': court, 'header': header})
+        next_url_name, next_url = request.GET.get('next', ''), ''
+        if next_url_name:
+            next_url = reverse(next_url_name, args=[cid])
+        return render(request, 'freports/delete_form.html', {'content': content, 'header': header, 'cancel_url': next_url})
     else:
         if request.POST.get('cancel_button'):
-            messages.warning(request, u"Видалення суду скасовано")
+            messages.warning(request, u"Видалення інформації про {} скасовано".format(court.name))
+            next_url = request.POST.get('cancel_next')
         elif request.POST.get('delete_button'):
-            court.delete()
-            messages.success(request, u"%s успішно видалений" % court)
+            judges = Judge.objects.filter(court_name=court)
+            if judges.count() > 0:
+                messages.error(request, u"Видалення неможливе. У складі даного суду наявні судді!")
+                next_url = request.POST.get('cancel_next')
+            else:
+                next_url = ''
+                court.delete()
+                messages.success(request, u"%s успішно видалений" % court)
 
-        return HttpResponseRedirect(reverse('courts_list'))
+        if next_url:
+            return HttpResponseRedirect(next_url)
+        else:
+            return HttpResponseRedirect(reverse('courts_list'))
 
 
 def valid_court(request_info):
